Add systemd support to cupsd (STR #3282)
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@11717 a1ca3aef-8c08-0410-bb20-df032aa958be

+++ scheduler/client.c
@@ -2516,7 +2516,7 @@ cupsdWriteClient(cupsd_client_t *con)	/* I - Client connection */
 
   bytes = (ssize_t)(sizeof(con->header) - (size_t)con->header_used);
 
-  if (!con->pipe_pid && bytes > httpGetRemaining(con->http))
+  if (!con->pipe_pid && bytes > (ssize_t)httpGetRemaining(con->http))
   {
    /*
     * Limit GET bytes to original size of file (STR #3265)...
